ref(profiling) Fix electron crash (#14216)
This PR fixes the electron crash observed in
https://github.com/getsentry/sentry-javascript/discussions/13978

I am not entirely sure as to why this causes a sigabrt, so I am working
around the issue (obtaining a coredump out of electron did not seem
trivial and my knowledge around electron isn't very extensive). The v8
options class and the constants are exposed correctly, I ruled that out,
however the crash still seems to happen when they are used in this
specific signature.

In order to have this running with electron, users will require to use
the electron/rebuild package, which is the recommended approach by
electron that rebuilds native node addons by providing the correct abi
headers for the electron version the user is running. For now, we wont
provide any prebuilt binaries and instead rely on the fallback mechanism
to load the correct module. I will reevaluate this if it causes issues
with bundling and look to add proper runtime electron detection.